+++ lib/rubocop/cop/rspec/file_path.rb
@@ -66,7 +66,7 @@ module RuboCop
         def camel_to_underscore(string)
           string.dup.tap do |result|
             result.gsub!(/([^A-Z])([A-Z]+)/, '\\1_\\2')
-            result.gsub!(/([A-Z])([A-Z][^A-Z]+)/, '\\1_\\2')
+            result.gsub!(/([A-Z])([A-Z\d][^A-Z\d]+)/, '\\1_\\2')
             result.downcase!
           end
         end